requirements.txt
fastapi==0.95.2
uvicorn==0.23.1
sqlalchemy==1.4.46
pydantic==1.10.9
python-multipart==0.0.6
jinja2
python-dotenv